+++ flup/data_structs/dyn_array.h
@@ -9,26 +9,26 @@
 #include "flup/tags.h"
 
 typedef struct flup_dyn_array {
-  unsigned int length;
-  unsigned int capacity;
+  size_t length;
+  size_t capacity;
   size_t elementSize;
   void* array;
 } flup_dyn_array;
 
 FLUP_ALLOCS_MEM
 FLUP_PUBLIC
-flup_dyn_array* flup_dyn_array_new(size_t elementSize, unsigned int capacity);
+flup_dyn_array* flup_dyn_array_new(size_t elementSize, size_t capacity);
 
 FLUP_ALLOCS_MEM
 FLUP_PUBLIC
-int flup_dyn_array_reserve(flup_dyn_array* self, unsigned int count);
+int flup_dyn_array_reserve(flup_dyn_array* self, size_t count);
 
 FLUP_DEALLOCS_MEM
 FLUP_PUBLIC
 void flup_dyn_array_free(flup_dyn_array* self);
 
 FLUP_PUBLIC
-int flup_dyn_array_get(flup_dyn_array* self, unsigned int index, void** element);
+int flup_dyn_array_get(flup_dyn_array* self, size_t index, void** element);
 
 FLUP_ALLOCS_MEM
 FLUP_PUBLIC
@@ -39,17 +39,17 @@ FLUP_PUBLIC
 int flup_dyn_array_prepend(flup_dyn_array* self, const void* element);
 
 FLUP_PUBLIC
-int flup_dyn_array_set(flup_dyn_array* self, unsigned int index, const void* element);
+int flup_dyn_array_set(flup_dyn_array* self, size_t index, const void* element);
 
 FLUP_DEALLOCS_MEM
 FLUP_PUBLIC
 int flup_dyn_array_trim(flup_dyn_array* self);
 
 FLUP_PUBLIC
-int flup_dyn_array_delete(flup_dyn_array* self, unsigned int index, unsigned int count);
+int flup_dyn_array_delete(flup_dyn_array* self, size_t index, size_t count);
 
 FLUP_PUBLIC
-int flup_dyn_array_insert(flup_dyn_array* self, unsigned int index, const void* element);
+int flup_dyn_array_insert(flup_dyn_array* self, size_t index, const void* element);
 
 #endif
